Added error handling for no joystick
Added Error Handling in case no joystick is connected.
Before this change, running 'HebiJoystick Input' was not possible without
a joystick connected and gave an error. Now, 'HebiJoystick Input' only gives
a warning if no joystick is connected and can be executed.

'HebiJoystick Input' still outputs data if no joystick is connected.
The reason for that is unclear, because the Level-2 Matlab S-Function for
'HebiJoystick Input' registers no 'Outputs' Method if no joystick is connected.
It would be better if no data would be published at all, but at the moment I don't
now if and how this is possible...